requirements.txt
shap==0.38.1
pandas==1.2.2
numpy==1.19.5
pillow==8.1.0
matplotlib==3.3.4
jupyter==1.0.0
streamlit==0.76.0
catboost==0.24.4
ppscore==1.2.0
pandas_profiling==2.11.0
seaborn==0.11.1
gplearn==0.4.1
